# lib/api-zod/src/generated/types/visaLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisaLeadBodyAnswers } from "./visaLeadBodyAnswers";
import type { VisaLeadBodyResult } from "./visaLeadBodyResult";

export interface VisaLeadBody {
  fullName: string;
  email: string;
  phone: string;
  partnerNationality: string;
  turnstileToken: string;
  score: number;
  result: VisaLeadBodyResult;
  answers: VisaLeadBodyAnswers;
  gclid?: string;
  utmSource?: string;
  utmCampaign?: string;
  utmMedium?: string;
  utmTerm?: string;
  utmContent?: string;
  referrer?: string;
}