# lib/api-zod/src/generated/types/willsLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WillsLeadBodyAnswers } from "./willsLeadBodyAnswers";
import type { WillsLeadBodyRoute } from "./willsLeadBodyRoute";

export interface WillsLeadBody {
  name: string;
  phone: string;
  email: string;
  turnstileToken: string;
  route: WillsLeadBodyRoute;
  answers: WillsLeadBodyAnswers;
  gclid?: string;
  utmSource?: string;
  utmCampaign?: string;
  utmMedium?: string;
  utmTerm?: string;
  utmContent?: string;
  referrer?: string;
}